feat(tracing): reuse the logs date range picker via a shared component
Extract the mature logs date-range picker into a shared, product-neutral
`DateRangePicker` (frontend/src/lib/components/DateFilter/DateRangePicker) and
migrate tracing onto it, replacing the generic `DateFilter`.

- The picker is decoupled from logs via props: `logicKey` (keys the backing
  logic so popover/custom-input/recent-history state and persisted history are
  isolated per surface), optional `dateOptions`, `allowedRollingDateOptions`,
  and optional `timezone`/`onTimezoneChange` (timezone selector hidden when
  absent, parsing falls back to UTC).
- `TimezoneSelect` moved to lib (it was generic).
- Logs `LogsDateRangePicker` becomes a thin wrapper; `LogsFilterBar` and the
  `NEW_LOGS_DATE_RANGE_PICKER` flag are untouched.
- Tracing gains the richer 11-preset list and persisted timezone state.

Generated-By: PostHog Code
Task-Id: cfb2ee66-54c7-4932-a425-594a31978a97
