+++ frontend/src/lib/components/DateFilter/DateRangePicker/index.ts
@@ -1,0 +1,3 @@
+export { DateRangePicker } from './DateRangePicker'
+export type { DateRangePickerProps } from './DateRangePicker'
+export { DEFAULT_DATE_RANGE_PICKER_OPTIONS } from './constants'
