+++ frontend/src/lib/components/TimezoneSelect/index.ts
@@ -1,0 +1,2 @@
+export { TimezoneSelect } from './TimezoneSelect'
+export type { TimezoneSelectProps } from './TimezoneSelect'
