+++ products/tracing/frontend/TracingFilterBar.tsx
@@ -12,8 +12,7 @@ import {
     LemonTag,
 } from '@posthog/lemon-ui'
 
-import { DateFilter } from 'lib/components/DateFilter/DateFilter'
-import { CUSTOM_OPTION_KEY } from 'lib/components/DateFilter/types'
+import { DateRangePicker } from 'lib/components/DateFilter/DateRangePicker'
 import { InfiniteSelectResults } from 'lib/components/TaxonomicFilter/InfiniteSelectResults'
 import { TaxonomicFilterSearchInput } from 'lib/components/TaxonomicFilter/TaxonomicFilter'
 import { taxonomicFilterLogic } from 'lib/components/TaxonomicFilter/taxonomicFilterLogic'
@@ -23,12 +22,10 @@ import { universalFiltersLogic } from 'lib/components/UniversalFilters/universal
 import { isUniversalGroupFilterLike } from 'lib/components/UniversalFilters/utils'
 import { dayjs } from 'lib/dayjs'
 import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
-import { DATE_TIME_FORMAT, formatDateRange } from 'lib/utils/datetime'
 
 import { DateRange } from '~/queries/schema/schema-general'
 import {
     AnyPropertyFilter,
-    DateMappingOption,
     FilterLogicalOperator,
     PropertyFilterType,
     PropertyOperator,
@@ -46,51 +43,11 @@ const taxonomicGroupTypes = [
     TaxonomicFilterGroupType.SpanResourceAttributes,
 ]
 
-const dateMapping: DateMappingOption[] = [
-    { key: CUSTOM_OPTION_KEY, values: [] },
-    {
-        key: 'Last 5 minutes',
-        values: ['-5M'],
-        getFormattedDate: (date: dayjs.Dayjs): string => date.subtract(5, 'minute').format(DATE_TIME_FORMAT),
-        defaultInterval: 'minute',
-    },
-    {
-        key: 'Last 30 minutes',
-        values: ['-30M'],
-        getFormattedDate: (date: dayjs.Dayjs): string => date.subtract(30, 'minute').format(DATE_TIME_FORMAT),
-        defaultInterval: 'minute',
-    },
-    {
-        key: 'Last 1 hour',
-        values: ['-1h'],
-        getFormattedDate: (date: dayjs.Dayjs): string => formatDateRange(date.subtract(1, 'h'), date.endOf('d')),
-        defaultInterval: 'hour',
-    },
-    {
-        key: 'Last 4 hours',
-        values: ['-4h'],
-        getFormattedDate: (date: dayjs.Dayjs): string => formatDateRange(date.subtract(4, 'h'), date.endOf('d')),
-        defaultInterval: 'hour',
-    },
-    {
-        key: 'Last 24 hours',
-        values: ['-24h'],
-        getFormattedDate: (date: dayjs.Dayjs): string => formatDateRange(date.subtract(24, 'h'), date.endOf('d')),
-        defaultInterval: 'hour',
-    },
-    {
-        key: 'Last 7 days',
-        values: ['-7d'],
-        getFormattedDate: (date: dayjs.Dayjs): string => formatDateRange(date.subtract(7, 'd'), date.endOf('d')),
-        defaultInterval: 'day',
-    },
-]
-
 export function TracingFilterBar(): JSX.Element {
     const { spansLoading } = useValues(tracingDataLogic())
     const { runQuery } = useActions(tracingDataLogic())
-    const { filters, utcDateRange } = useValues(tracingFiltersLogic())
-    const { setDateRange, setServiceNames, setFilterGroup, setViewMode, setCompareMode } =
+    const { filters, utcDateRange, timezone } = useValues(tracingFiltersLogic())
+    const { setDateRange, setTimezone, setServiceNames, setFilterGroup, setViewMode, setCompareMode } =
         useActions(tracingFiltersLogic())
     const { dateRange, serviceNames, filterGroup, viewMode, compareMode } = filters
 
@@ -125,18 +82,12 @@ export function TracingFilterBar(): JSX.Element {
                                     })
                                 }}
                             />
-                            <DateFilter
-                                size="small"
-                                dateFrom={dateRange.date_from}
-                                dateTo={dateRange.date_to}
-                                dateOptions={dateMapping}
-                                onChange={(changedDateFrom, changedDateTo) => {
-                                    setDateRange({ date_from: changedDateFrom, date_to: changedDateTo })
-                                }}
-                                allowTimePrecision
-                                allowFixedRangeWithTime
-                                allowedRollingDateOptions={['minutes', 'hours', 'days', 'weeks', 'months']}
-                                use24HourFormat
+                            <DateRangePicker
+                                logicKey="tracing"
+                                dateRange={dateRange}
+                                setDateRange={setDateRange}
+                                timezone={timezone}
+                                onTimezoneChange={setTimezone}
                             />
                             <LemonButton
                                 size="small"
